Karishma added changes.

+++ docsapi/index.js
@@ -22,6 +22,8 @@ app.use((req, res, next) => {
   }
   next();
 });
+
+// console.log('globalVariable',globalVariable)
  
 //create database connection
 const conn = mysql.createConnection({
@@ -190,6 +192,7 @@ app.get('/api/get_serviceForm/:id',(req, res) => {
     if(err){
       res.send(JSON.stringify({"status": 500, "error": err, "response": null}))
     }else{
+      // console.log('results',results)
       if(results&&results.length!=0){
         var newArr = []
         for (var i = 0; i < results.length; i++) {
@@ -198,8 +201,9 @@ app.get('/api/get_serviceForm/:id',(req, res) => {
             if(err){
               res.send(JSON.stringify({"status": 500, "error": err, "response": null}))
             }else{
+              // console.log('result[0]',result[0])
               newArr.push(result[0])
-              if(i>=results.length){
+              if(newArr.length===results.length){
                 res.send(JSON.stringify({"status": 200, "error": null, "response": newArr}));
               }
             }
@@ -218,31 +222,29 @@ app.get('/api/get_serviceForm',(req, res) => {
   let query = conn.query(sql, (err, results) => {
     if(err){
       res.send(JSON.stringify({"status": 500, "error": err, "response": null}))
-    }else{
+    }else{ 
       if(results&&results.length!=0){
-        var newArr = []
-        for (var i = 0; i < results.length; i++) {
-          var serviceName = ''
-          var mandatory = results[i].IS_MANDATORY
-          var serviceId = results[i].SERVICE_ID
-          let sqlService = "SELECT NAME FROM SERVICE WHERE id="+results[i].SERVICE_ID;
-          let queryService = conn.query(sqlService, (err, result) => {
-            if(err){
-            }else{
-              serviceName = result[0].NAME
-            }
-          })
-          let sql = "SELECT * FROM FORM_FIELDS WHERE id="+results[i].FORM_FIELD_ID;
-          let query = conn.query(sql, (err, result) => {
-            if(err){
-              res.send(JSON.stringify({"status": 500, "error": err, "response": null}))
-            }else{
-              newArr.push({data:result[0],mandatory:mandatory,serviceName:serviceName,serviceId:serviceId})
-              if(i>=results.length){
-                res.send(JSON.stringify({"status": 200, "error": null, "response": newArr}));
+        getServiceName();
+        async function getServiceName(){
+          var newArr = []
+          for (var i = 0; i < results.length; i++) {
+            var mandatory = results[i].IS_MANDATORY
+            var serviceId = results[i].SERVICE_ID
+            var formFeildId = results[i].FORM_FIELD_ID
+            var serviceName = await fetch_serviceName(serviceId);
+            var prevServiceId = serviceId;
+            let sql = "SELECT * FROM FORM_FIELDS WHERE id="+formFeildId;
+            let query = conn.query(sql, (err, result) => {
+              if(err){
+                res.send(JSON.stringify({"status": 500, "error": err, "response": null}))
+              }else{
+                newArr.push({data:result[0],mandatory:mandatory,serviceName:serviceName,serviceId:prevServiceId})
+                if(newArr.length===results.length){
+                  res.send(JSON.stringify({"status": 200, "error": null, "response": newArr}));
+                }
               }
-            }
-          })
+            })
+          }
         }
       }else{
         res.send(JSON.stringify({"status": 404, "error": null, "response": {message: 'Data not found.'}}));
@@ -251,11 +253,26 @@ app.get('/api/get_serviceForm',(req, res) => {
   })
 });
 
-//update service form fields
+function fetch_serviceName(id){
+  return new Promise(function(resolve,reject){
+    let sqlService = "SELECT NAME FROM SERVICE WHERE id="+id;
+    let queryService = conn.query(sqlService, (err, result) => {
+      if(err){
+        console.log(err);
+        reject(err);
+      }else{
+        serviceName = result[0].NAME
+        resolve(serviceName); 
+      }
+    })
+  })
+}// end function
+
+//update service form fields(not working)
 app.put('/api/update_serviceForm/:id',(req, res) => {
   let sql = "UPDATE FORM_FIELDS SET name='"+req.body.fieldName+"', label='"+req.body.label+
   "', type='"+req.body.type+"', option='"+req.body.option+"', id='"+req.body.fieldId+
-  "', sequenceno='"+req.body.sequenceno+"' WHERE id="+req.params.id;
+  "', sequenceno='"+req.body.sequenceno+"' WHERE id='"+req.params.id;
   let query = conn.query(sql, (err, results) => {
     if(err){
       res.send(JSON.stringify({"status": 500, "error": err, "response": null}));
@@ -273,14 +290,28 @@ app.put('/api/update_serviceForm/:id',(req, res) => {
   })
 });
 
-//delete service form fields
+function get_formFieldDeleteRes(id){
+  return new Promise(function(resolve,reject){
+    let sql = "DELETE FROM SERVICE_FORM_FIELD WHERE service_id="+id;
+    let query = conn.query(sql, (err, results) => {
+      if(err){
+        console.log(err);
+        reject(err);
+      }else{
+        resolve(true); 
+      }
+    })
+  })
+}// end function
+
+//delete service form fields (not working)
 app.delete('/api/delete_serviceForm/:id',(req, res) => {
-  let sql = "DELETE FROM FORM_FIELDS WHERE id="+req.params.id;
-  let query = conn.query(sql, (err, results) => {
-    if(err){
-      res.send(JSON.stringify({"status": 500, "error": err, "response": null}));
-    }else{
-      let sql = "DELETE FROM SERVICE_FORM_FIELD WHERE service_id="+req.params.id;
+  getFormDeleteResult();
+  async function getFormDeleteResult(){
+    var isFormDelete = await get_formFieldDeleteRes(req.params.id)
+    // console.log('isFormDelete',isFormDelete)
+    if(isFormDelete){
+      let sql = "DELETE FROM FORM_FIELDS WHERE id="+req.params.id;
       let query = conn.query(sql, (err, results) => {
         if(err){
           res.send(JSON.stringify({"status": 500, "error": err, "response": null}));
@@ -289,7 +320,22 @@ app.delete('/api/delete_serviceForm/:id',(req, res) => {
         }
       })
     }
-  })
+  }
+  // let sql = "DELETE FROM FORM_FIELDS WHERE id="+req.params.id;
+  // let query = conn.query(sql, (err, results) => {
+  //   if(err){
+  //     res.send(JSON.stringify({"status": 500, "error": err, "response": null}));
+  //   }else{
+  //     let sql = "DELETE FROM SERVICE_FORM_FIELD WHERE service_id="+req.params.id;
+  //     let query = conn.query(sql, (err, results) => {
+  //       if(err){
+  //         res.send(JSON.stringify({"status": 500, "error": err, "response": null}));
+  //       }else{
+  //         res.send(JSON.stringify({"status": 200, "error": null, "response": results}));
+  //       }
+  //     })
+  //   }
+  // })
 });
 
 //add service documents
@@ -335,7 +381,7 @@ app.get('/api/get_serviceDocument/:id',(req, res) => {
               res.send(JSON.stringify({"status": 500, "error": err, "response": null}))
             }else{
               newArr.push(result[0])
-              if(i>=results.length){
+              if(newArr.length===results.length){
                 res.send(JSON.stringify({"status": 200, "error": null, "response": newArr}));
               }
             }
@@ -344,6 +390,33 @@ app.get('/api/get_serviceDocument/:id',(req, res) => {
       }else{
         res.send(JSON.stringify({"status": 404, "error": null, "response": {message: 'Data not found.'}}));
       }
+    }
+  })
+});
+
+//add service data
+app.post('/api/add_serviceFormData',(req, res) => {
+  let sql = "SELECT order_id FROM ORDER_SERVICE_DATA ORDER BY order_id DESC";
+  let query = conn.query(sql, (err, results) => {
+    if(err){
+      res.send(JSON.stringify({"status": 500, "error": err, "response": null}))
+    }else{
+      if(results&&results.length!=0){
+        var id = parseInt(results[0].order_id)+1
+      }else{
+        var id = 1
+      }  
+      let data = {order_id:id,service_id:req.body.serviceId,field_name:req.body.fieldName,field_value:req.body.fieldValue};
+      // let sql = "INSERT INTO ORDER_SERVICE_DATA SET ?";
+      // let query = conn.query(sql, data,(err, results) => {
+      //   if(err) res.send(JSON.stringify({"status": 500, "error": err, "response": null}));;
+      //   let data = {service_id:req.body.serviceId,document_id:id,is_mandatory: req.body.mandatoryField};
+      //   let sql = "INSERT INTO DOCUMENT_SERVICE SET ?";
+      //   let query = conn.query(sql, data,(err, results) => {
+      //     if(err) res.send(JSON.stringify({"status": 500, "error": err, "response": null}));
+      //     res.send(JSON.stringify({"status": 200, "error": null, "response": results}));
+      //   });
+      // });
     }
   })
 });
